Refactor language
Currently, `language` is an enum field duplicated in both `Story` and `NostrUser` which
make it harder to maintain and allowing more languages.

This big refactoring changes the behavior by removing both enums and relying only on the `NostrUser`
associated to a `Story` to controle the adventure language.

The gem `i18n_data` helps to display languages list in `NostrUser` form select options.

When an adventure is triggered from the frontend interface, user has to choose the nostr user that
would publish the story in the corresponding record language.

Design has also been enhanced a bit with additional simple_form i18n translations.