Prevent updating nostr account language once it has been created
When a nostr account has been created, it should not be possible to update
its language to avoid users to subscribe to the account and have ununderstandable
stories appearing.

It also prevent messing up IA with a dropper story started in a language and ended in
another if language is changed in a middle of a story.