Refactor header navigation using Flowbite
Fixes #93

I wanted to avoid Flowbite javascript as much as possible to foster CSS only solution.
It worked for a while but menu has become harder to maintain to deal with main menu items,
dropdown submenu and dropdown logged in user menu.

Because of it, we should give Flowbite a new chance and get a mobile friendly responsive header
experience to handle all of this dropdown data.

Details:
- Replace previous CSS only header by Flowbite version
- Extract CSS from main Tailwind to dedicated components/plugins files
- Move story related menu items as main menu dropdown