Add `ruby-lsp-rails` gem to improve rich features in editors
Adding the [ruby-lsp-rails](https://github.com/Shopify/ruby-lsp-rails) gem to the project
allows to get a better understanding of methods and Rails code inside the editor showing
as a hover popup (for example visualize a model attributes with type).

Rails 8 planned to make it integrated into the next major version so let's experiment it
until it reach the big release.